+++ src/middlewares/error-handler.ts
@@ -2,6 +2,7 @@ import { GeneralError } from '../classes/general-error';
 import { HTTP_CODES } from '../constants/http-codes';
 import { errorResponse } from '../services/error-response';
 import { ErrorRequestHandler, NextFunction, Request, Response } from 'express';
+import Joi from 'joi';
 
 export const errorHandler: ErrorRequestHandler = async (
   error: Error,
@@ -10,14 +11,16 @@ export const errorHandler: ErrorRequestHandler = async (
   next: NextFunction
 ): Promise<void> => {
   try {
+    const { ValidationError } = Joi;
     console.error(
       `detailed information about the error in the service: `,
       error
     );
     const err: GeneralError = errorResponse(error);
     const response = err.buildResponse();
+    let codeError = error instanceof ValidationError ? 400 : err.code;
 
-    res.status(err.code).json(response);
+    res.status(codeError).json(response);
   } catch (error) {
     res.status(HTTP_CODES.ServerError).json(error);
   }