+++ CHANGELOG.md
@@ -6,3 +6,9 @@
 - Add base configuration
 - Add typeorm migrations
 - Add typeorm entities
+
+### 0.2.1
+
+- user entity crud created
+- documentation for user entity
+- fix error joi validation